+++ modules/front-end/src/app/features/safe/organizations/project/project.component.ts
@@ -5,12 +5,11 @@ import { OrganizationService } from '@services/organization.service';
 import { EnvService } from '@services/env.service';
 import { NzMessageService } from "ng-zorro-antd/message";
 import {PermissionsService} from "@services/permissions.service";
-import {generalResourceRNPattern, permissionActions} from "@shared/permissions";
 import {MessageQueueService} from "@services/message-queue.service";
 import { FormBuilder, FormGroup, Validators } from "@angular/forms";
 import { EnvSecretService } from "@services/env-secret.service";
 import { copyToClipboard } from '@utils/index';
-import { ResourceTypeEnum } from "@shared/policy";
+import { ResourceTypeEnum, generalResourceRNPattern, permissionActions } from "@shared/policy";
 
 @Component({
   selector: 'app-project',
